delete formatting file and merge augmenting with process_data.py
